# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación de la API",
        "description": "Esta es la documentación generada automáticamente por Swagger usando anotaciones OpenAPI",
        "contact": {
            "email": "tucorreo@dominio.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor principal"
        }
    ],
    "paths": {
        "/api/estudiantes": {
            "get": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Listar todos los estudiantes",
                "operationId": "209e318b7ac388b93bccf7af4162e33b",
                "responses": {
                    "200": {
                        "description": "Lista de estudiantes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Crear un nuevo estudiante",
                "operationId": "bf7df90391b947c8dc2f77570e61358a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "cedula",
                                    "correo",
                                    "paralelo_id"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "cedula": {
                                        "type": "string"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "paralelo_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Estudiante creado"
                    }
                }
            }
        },
        "/api/estudiantes/{id}": {
            "get": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Obtener un estudiante por ID",
                "operationId": "fcd2a65dde4ddb1efd0765a95b43b9fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del estudiante"
                    },
                    "404": {
                        "description": "Estudiante no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Actualizar un estudiante",
                "operationId": "a0d23067013987f1a070c064758e522b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "cedula": {
                                        "type": "string"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "paralelo_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Estudiante actualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Estudiantes"
                ],
                "summary": "Eliminar un estudiante",
                "operationId": "caeee49dd1fc2f0619a1069d2e414f0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Estudiante eliminado"
                    }
                }
            }
        },
        "/api/paralelos": {
            "get": {
                "tags": [
                    "Paralelos"
                ],
                "summary": "Listar todos los paralelos",
                "operationId": "30409aa4e072067e1395716421b89f6c",
                "responses": {
                    "200": {
                        "description": "Lista de paralelos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Paralelos"
                ],
                "summary": "Crear un nuevo paralelo",
                "operationId": "63f2e6de83fe35723b5077a59ec0f619",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paralelo creado exitosamente"
                    }
                }
            }
        },
        "/api/paralelos/{id}": {
            "get": {
                "tags": [
                    "Paralelos"
                ],
                "summary": "Obtener un paralelo por ID",
                "operationId": "4d4a75bdf60cae9e6845622d9d5089d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paralelo encontrado"
                    },
                    "404": {
                        "description": "Paralelo no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Paralelos"
                ],
                "summary": "Actualizar un paralelo existente",
                "operationId": "09c0363579e6f12d0cf5d6a8934f4fad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Paralelo actualizado"
                    },
                    "404": {
                        "description": "Paralelo no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Paralelos"
                ],
                "summary": "Eliminar un paralelo por ID",
                "operationId": "e1db2fb826fbe8f0eb7a8c73c9222f49",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Paralelo eliminado"
                    },
                    "404": {
                        "description": "Paralelo no encontrado"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Estudiantes",
            "description": "Operaciones relacionadas con los estudiantes"
        },
        {
            "name": "Paralelos",
            "description": "Operaciones relacionadas a los paralelos"
        }
    ]
}